% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exercise.R
\name{check_exercise}
\alias{check_exercise}
\title{Check/Test solution for a specific exercise}
\usage{
check_exercise(dir, exercise_id, exercise_hash)
}
\arguments{
\item{dir}{character, folder containing exercises and user uploads.}

\item{exercise_id}{integer, ID of the exercise (as in database).}

\item{exercise_hash}{string, combination of random hash, assignment-timestamp, and user ID
(see database). Used to identify a specific exercise for a specific user.}
}
\value{
No explicit return, creates output on stdout.
}
\description{
TODO
}
