% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_compile.R
\name{exr_get_content_by_id}
\alias{exr_get_content_by_id}
\title{Get the content of an xml node with a specific id}
\usage{
exr_get_content_by_id(x, id, elem = "div")
}
\arguments{
\item{x}{\code{xml_node} object}

\item{id}{\code{character}, ID of the element}

\item{elem}{\code{character}, default \code{"div"}}
}
\value{
Returns the content of this specific element given
the element (default \code{"div"}) with the specific ElementID
as specified on input \code{id}.

Returns the xml/html node given input \code{elem} (default \code{"div"})
and \code{id}. If input \code{id = NA}, \code{NA} will be returned.
}
\description{
Helper function for \code{\link[exerciser]{exr_compile}}.
}
\author{
Reto Stauffer
}
