% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_compile.R
\name{exr_get_level1_id}
\alias{exr_get_level1_id}
\title{Extract ElementID from html file}
\usage{
exr_get_level1_id(doc, what, required = TRUE, verbose = FALSE)
}
\arguments{
\item{doc}{\code{xml_node} object}

\item{what}{character, name of the IDs to be found. Used
in a regular expression (see Details)}

\item{required}{logical}
}
\value{
Returns \code{NA} or a \code{character}, see Details.
}
\description{
Helper function for \code{\link[exerciser]{exr_compile}}.
}
\details{
Used to find specific elements in the xml document.
Extracts all \code{//*/div[@class='section level1']} IDs matching
the pattern \code{^<what>}. Thus, if \code{what = "description"}
we will find the entry \code{id="description"} but also further
elements (e.g., \code{id="description-1"}, \code{id="description-2"}).

The following can happen:
\itemize{
    \item No matching entry, but \code{required = TRUE}: raise an error.
    \item No matching entry, but \code{required = FALSE}: NA.
    \item Multiple matching entries found: raise an error.
    \item Exactely one matching entry found: return character, ID of the element.
}
}
\author{
Reto Stauffer
}
