% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_compile.R
\name{exr_parse_settings}
\alias{exr_parse_settings}
\title{Parse settings section of a rendered exerciseR exercise}
\usage{
exr_parse_settings(doc, id, verbose = FALSE, elem = "div")
}
\arguments{
\item{doc}{\code{xml_node} object}

\item{id}{\code{character}, ElementID}

\item{verbose}{\code{logical}}

\item{elem}{\code{character}, default \code{"div"}}
}
\value{
Returns a list with character vectors for elements
\code{blacklist}, \code{whitelist}, and \code{files}, and a numeric
value (or numeric \code{NA}) for the \code{points}.
}
\description{
Helper function for \code{\link[exerciser]{exr_compile}}.
Extracts the information from the settings section of the Rmd/html file.
}
\details{
We are expecting that the settings section of an exerciseR
exercise contains the following items:
\itemize{
   \item blacklist: comma separated list of commands which are not allowed to
         be used by the user (in addition to the default ones).
   \item whitelist: comma separated list of commands which can be used, allows
         to overrule the default blacklist (and can be used to give the user
         a hint what should be used). However, not mandatory to be used.
   \item files: list of files which come along with the exercise. If present,
         these files will be automatically added when executing user code and
         the UI will prohibit the users to upload a file with one of these names
         not to get in conflict.
   \item points: numeric value, points for this exercise.
}
Note that all entries are optional!
}
\author{
Reto Stauffer
}
