% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_compile.R
\name{exr_parse_tests}
\alias{exr_parse_tests}
\title{Parse test section of a rendered exerciseR exercise}
\usage{
exr_parse_tests(doc, id, verbose = TRUE, elem = "div")
}
\arguments{
\item{doc}{\code{xml_node} object}

\item{id}{\code{character}, ElementID}

\item{verbose}{\code{logical}}

\item{elem}{\code{character}, default \code{"div"}}
}
\value{
Returns a list with all tests in the same order as defined
in the Rmd file. The name of the list elements define whether or not
a test is visible or not.
}
\description{
Helper function for \code{\link[exerciser]{exr_compile}}.
Extracts the information from the tests section.
}
\details{
The test section can look as follows in the markdown file:
 ----------------------------------
 # Tests
 * expect_true(...)
 * expect_identical(...)
 ## Visible
 * expect_true(...)
 ```
 <some custom test code>
 ```
 ## Invisible
 * expect_true(...)
 ```
 <some custom test code>
 ```
 ----------------------------------
 
 Tests not defined in a subsection (## Visible or ## Invisible)
 will be treated as "invisible" by default!
 Those in ## Visible will be visible to the users. Those in
 ## Invisible will not be shown on the UI.
 Tests will be executed in the sequence as shown above. Multiple
 Subsections are allowed (e.g., ## Visible followed by ## Invisible
 followed by another ## Visible subsection).
}
\author{
Reto Stauffer
}
